requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
cloudpickle==3.0.0
colorama==0.4.6
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.14.0
idna==3.10
jmespath==1.0.1
joblib==1.4.2
numpy==2.0.2
pandas==2.2.3
pydantic==2.11.3
pydantic_core==2.33.1
pymongo==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
threadpoolctl==3.6.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
